% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_plot_utils.R
\name{facet_violin}
\alias{facet_violin}
\title{分面小提琴图}
\usage{
facet_violin(data, vars, groupby)
}
\arguments{
\item{data}{数据框}

\item{vars}{指标向量}

\item{groupby}{分组变量名}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   ggplot 对象
}\if{html}{\out{</div>}}
}
\description{
分面小提琴图
}
