% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_clinkit_opt}
\alias{get_clinkit_opt}
\title{Get Global Options for ClinKit}
\usage{
get_clinkit_opt(opt_name, default)
}
\arguments{
\item{opt_name}{Character scalar. The name of the option to retrieve (without the "ClinKit." prefix).}

\item{default}{The default value to return if the option is not found.}
}
\value{
The value of the specified global option, or the default value if it is not set.
}
\description{
Retrieves package-specific global options. If the user has not set a specific
option globally (using \code{options()}), the function returns a provided default value.
}
\keyword{internal}
