% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_forest.R
\name{get_forest_theme}
\alias{get_forest_theme}
\alias{tm_blue}
\alias{tm_green}
\alias{tm_cyan}
\alias{tm_default}
\title{Get Predefined Forest Plot Themes}
\usage{
get_forest_theme(
  theme_name = c("blue", "green", "cyan", "default"),
  base_size = 12,
  background_levels = NULL
)

tm_blue(base_size = 12, background_levels = NULL)

tm_green(base_size = 12, background_levels = NULL)

tm_cyan(base_size = 12, background_levels = NULL)

tm_default(base_size = 12, background_levels = NULL)
}
\arguments{
\item{theme_name}{Theme name: "blue", "green", "cyan", or "default"}

\item{base_size}{Base font size, default 12}

\item{background_levels}{Background color levels vector for subgroup hierarchy}
}
\value{
Forest plot theme object
}
\description{
Provides several predefined color themes for forest plots with Times New Roman font.
}
\examples{
\dontrun{
tm <- get_forest_theme("blue")
result <- subgroup_forest(..., tm = tm)
}
}
