% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_baseline.R
\name{is_normal_overall}
\alias{is_normal_overall}
\title{Check if all groups for a variable are normal}
\usage{
is_normal_overall(normality_results, var)
}
\arguments{
\item{normality_results}{Data frame from perform_normality_tests}

\item{var}{Character, variable name to check}
}
\value{
Logical, TRUE if all groups are normal, FALSE otherwise
}
\description{
Check if all groups for a variable are normal
}
