% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_baseline.R
\name{print_normality_summary}
\alias{print_normality_summary}
\title{Print summary of normality test results}
\usage{
print_normality_summary(normality_results)
}
\arguments{
\item{normality_results}{Data frame from perform_normality_tests}
}
\description{
Print summary of normality test results
}
