% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_univariate.R
\name{save_analysis_result}
\alias{save_analysis_result}
\title{Save Analysis Results (Internal Function)}
\usage{
save_analysis_result(result, output_dir, format, predictor, outcome_name)
}
\description{
Save Analysis Results (Internal Function)
}
\keyword{internal}
