% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_interaction.R
\name{save_highlow_table}
\alias{save_highlow_table}
\title{Save High-Low Analysis Table (Internal)}
\usage{
save_highlow_table(results, file_path, title)
}
\description{
Save High-Low Analysis Table (Internal)
}
\keyword{internal}
