% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_interaction.R
\name{save_interaction_table}
\alias{save_interaction_table}
\title{Save Interaction Table (Internal)}
\usage{
save_interaction_table(interaction_result, file_path)
}
\description{
Save Interaction Table (Internal)
}
\keyword{internal}
