% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_quartile.R
\name{save_multinomial_quartile_table}
\alias{save_multinomial_quartile_table}
\title{Save Multinomial Quartile Results Table (Internal)}
\usage{
save_multinomial_quartile_table(results, file_path)
}
\description{
Save Multinomial Quartile Results Table (Internal)
}
\keyword{internal}
