% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_multivariable.R
\name{save_multinomial_table}
\alias{save_multinomial_table}
\title{Save Multinomial Results Table (Internal)}
\usage{
save_multinomial_table(results, file_path)
}
\description{
Save Multinomial Results Table (Internal)
}
\keyword{internal}
