% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_forest.R
\name{setup_font_config}
\alias{setup_font_config}
\title{Setup Font Configuration for Math Symbols (Internal)}
\usage{
setup_font_config(math_font = NULL, ensure_italic_p = TRUE)
}
\description{
Setup Font Configuration for Math Symbols (Internal)
}
\keyword{internal}
