% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_clinkit_project}
\alias{use_clinkit_project}
\title{Initialize a ClinKit Analysis Project}
\usage{
use_clinkit_project(path = ".")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Creates the core directory structure and the master configuration file
for a standardized clinical research workflow.
}
