% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_baseline}
\alias{use_module_baseline}
\title{Generate Baseline Table Module Script}
\usage{
use_module_baseline(path = "scripts/01_baseline_table.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Baseline Table Module Script
}
