% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_forest}
\alias{use_module_forest}
\title{Generate Subgroup Forest Plot Module Script}
\usage{
use_module_forest(path = "scripts/05_subgroup_forest.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Subgroup Forest Plot Module Script
}
