% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_highlow}
\alias{use_module_highlow}
\title{Generate High-Low Joint Analysis Module Script}
\usage{
use_module_highlow(path = "scripts/07_highlow_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate High-Low Joint Analysis Module Script
}
