% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_multivariable}
\alias{use_module_multivariable}
\title{Generate Multivariable Analysis Module Script}
\usage{
use_module_multivariable(path = "scripts/03_multivariable_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Multivariable Analysis Module Script
}
