% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_nri}
\alias{use_module_nri}
\title{Generate Reclassification Analysis Module Script}
\usage{
use_module_nri(path = "scripts/06_nri_idi_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Reclassification Analysis Module Script
}
