% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_rcs}
\alias{use_module_rcs}
\title{Generate RCS Plot Module Script}
\usage{
use_module_rcs(path = "scripts/04_rcs_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate RCS Plot Module Script
}
