% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_reclassification}
\alias{use_module_reclassification}
\title{Generate Reclassification Analysis Module Script}
\usage{
use_module_reclassification(path = "scripts/06_reclassification_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Reclassification Analysis Module Script
}
