% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scaffolds.R
\name{use_module_univariate}
\alias{use_module_univariate}
\title{Generate Univariate Analysis Module Script}
\usage{
use_module_univariate(path = "scripts/02_univariate_analysis.R")
}
\arguments{
\item{path}{Character. Target file path for the generated script.}
}
\description{
Generate Univariate Analysis Module Script
}
