% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_plot_utils.R
\name{facet_violin}
\alias{facet_violin}
\title{Faceted Violin Plot}
\usage{
facet_violin(data, vars, groupby)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Character vector, names of indicator variables to plot}

\item{groupby}{Character scalar, name of grouping variable}
}
\value{
ggplot object
}
\description{
Draws faceted violin plots for multiple indicators, grouped by a grouping variable,
with each indicator in a separate facet
}
\details{
This function will:
\enumerate{
\item Select specified columns and convert to long format
\item Convert grouping variable to factor
\item Draw violin plots with boxplots inside, faceted by indicator
}
}
\examples{
\dontrun{
p <- facet_violin(iris, c("Sepal.Length", "Sepal.Width"), "Species")
print(p)
}

}
