% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rcs.R
\name{generate_rcs_plot}
\alias{generate_rcs_plot}
\title{RCS 逻辑回归 + 图}
\usage{
generate_rcs_plot(
  outcome,
  predictor,
  data,
  output_folder,
  models = covariates,
  use_model = FALSE,
  filename = NULL
)
}
\arguments{
\item{outcome}{结局变量名}

\item{predictor}{连续预测变量名}

\item{data}{数据框}

\item{output_folder}{输出目录}

\item{models}{协变量向量，默认 NULL}

\item{use_model}{是否加协变量，默认 FALSE}

\item{filename}{自定义文件名，默认 NULL（自动生成）}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        列表：plot + 非线性/总体 P 值（字符 + 数值）
}\if{html}{\out{</div>}}
}
\description{
RCS 逻辑回归 + 图
}
\examples{
\dontrun{
generate_rcs_plot("death", "SIRI", data, "./RCS")
}
}
