% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rcs.R
\name{run_analysis_rcs}
\alias{run_analysis_rcs}
\title{批量 RCS 分析}
\usage{
run_analysis_rcs(
  data,
  related_index,
  all_outcomes,
  folder_path,
  outcomes_map,
  covariates
)
}
\arguments{
\item{data}{数据框}

\item{related_index}{预测变量向量}

\item{all_outcomes}{结局变量向量}

\item{folder_path}{根目录}

\item{outcomes_map}{结局中文映射 named vector/list}

\item{covariates}{协变量向量}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        data.frame：predictor / outcome / model_type / P 值
}\if{html}{\out{</div>}}
}
\description{
批量 RCS 分析
}
\examples{
\dontrun{
run_analysis_rcs(data, c("SIRI"), c("death"), "./RCS", outcomes_map, covariates)
}
}
