% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_multivariable.R
\name{run_multivariable_logistic_regression}
\alias{run_multivariable_logistic_regression}
\title{多因素逻辑回归（4 层模型）}
\usage{
run_multivariable_logistic_regression(
  related_index,
  all_outcomes,
  data,
  folder_path,
  outcomes_map,
  models_list
)
}
\arguments{
\item{related_index}{预测变量}

\item{all_outcomes}{结局变量}

\item{data}{数据框}

\item{folder_path}{结果保存目录（每个预测变量×结局生成一个txt)}

\item{outcomes_map}{结局变量中文名映射（list）}

\item{models_list}{协变量列表，长度 3（model2、3、4）}
}
\value{
包含预测变量×结局的多因素逻辑回归结果的dataframe
names 必须与 model2、model3、model4 对应
}
\description{
多因素逻辑回归（4 层模型）
}
