% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_univariate.R
\name{run_univariate_logistic_regression}
\alias{run_univariate_logistic_regression}
\title{单因素 Logistic}
\usage{
run_univariate_logistic_regression(
  related_index,
  all_outcomes,
  data,
  folder_path,
  outcomes_map
)
}
\arguments{
\item{related_index}{预测变量}

\item{all_outcomes}{结局变量}

\item{data}{数据框}

\item{folder_path}{结果保存目录}

\item{outcomes_map}{中文映射}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         data.frame（OR / 95\%CI / p）
}\if{html}{\out{</div>}}
}
\description{
单因素 Logistic
}
